#38
- preset/node/node: перенесен в preset/node (major)
- preset/node/handler: перенесен в preset/node-handler (major)
- package.json: /providers и /utils теперь недоступны (major)
